requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-dotenv==1.0.0
supabase==2.16.0
psycopg2-binary==2.9.9
anyio==3.7.1
pyjwt==2.8.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0